Modified borrow process
Earlier a copy record would get deleted everytime a book was borrowed.
Now after adding a copy_id(foreign key referencing pk of bookInstance table) column to the borrow table
a copy is being borrowed instead of a book. But now need to solve the issue of the same copy being borrowed
multiple times. This is happening because previously the first record was being retrieved. (Line 162 of views.py)